+++ lib/src/models/content.dart
@@ -1,8 +1,6 @@
 import 'package:powerbook/mixins.dart';
 
 /// A generic class that represents a single piece of content
-abstract class Content with Renderable, Indexable {
-  final String type;
-
-  Content({required this.type});
+abstract class Content with Renderable {
+  Content();
 }