+++ test/mocks/mock_content.dart
@@ -1,20 +1,10 @@
 import 'package:powerbook/models.dart';
 
 class MockContent extends Content {
-  MockContent() : super(type: "mockContent");
+  MockContent();
 
   @override
   String toHtml() {
     throw UnimplementedError();
-  }
-
-  @override
-  String getId() {
-    return "mocked-$hashCode";
-  }
-
-  @override
-  String getTitle() {
-    return "Mocked Content";
   }
 }